20250730 很多修改
加入电子书处理API (roj.ebook)
	修复NovelFrame默认正则问题
	修复EpubWriter的XML转义问题
	修复XML解析问题
加入Roj234 Archive(roar)的读写 (roj.archive.roar)
	QZArchiver进入稳定API
加入基于FPE（Format Preserve Encryption）的随机不重复序列生成器. aka. 数组采样器
	优化部分密码学算法性能
加入基于Tracker的DDNS服务实现
加入基于DP的模糊文本搜索算法。
加入Javap（RojASM based）调试工具
Panger支持依赖注入
加入属性以完全禁用Terminal功能（不带fallback而是抛出异常）
加入了roj.config.auto.Comment注解，现在序列化的POJO支持注释了
加入了MC数据包辅助工具
(Config) setProperty now use TypedKey

收紧反射权限
名称标准化
	Reflection等

bug修复：
	打开空zip时的NPE
	FPM会创建很多PHP进程
	修复流控算法的一些bug (还有几个)
	撤回之前PrefixDelegation对斜杠的处理
	被close的进度条依然渲染
	优化Mapper文件大小
	fix ParamNameMapper现在是完全根据参数位置，而不是slot选择了
	修复der序列化问题，加入v2签名，修复之前证书选择的bug，优化EdDSA参数名称，禁用自己写的X25519DH
	修复序列化器不支持Boolean
	修复编译器强制要求文本块缩进
	修复MIXIN注解处理器的bug
	修复Lavac的类型转换和一些消息问题
	修复生成器函数的问题
	Fixed a critical problem cause TypeParam#toDesc throw NPE
	修复了CMap#getBool不取默认值的问题
	修复了QOI编解码在最后数个像素时偶发的索引越界
	修复了日期解析的索引越界
	ByteList现在能正确的将缓冲区扩展到最大容量了
	修复了BT处理的一些bug

Lavac:
	Lavac.NaE现在携带位置信息

FMD:
	加入了MappingBuilder以波兰表达式自动化生成映射
	支持项目和工作空间的导入导出
	自动生成的指令帮助现在可以部分修改了
	注解缓存白名单
	加入鼠标事件处理

+++ java/roj/archive/zip/EntryMod.java
@@ -10,7 +10,7 @@ public final class EntryMod {
 	String name;
 	ZEntry entry;
 
-	public static final int KEEP_TIME = 1, UFS = 2, LARGE = 4, COMPRESS = 8;
+	public static final byte KEEP_TIME = 1, UFS = 2, LARGE = 4, COMPRESS = 8;
 	@MagicConstant(flags = {KEEP_TIME, UFS, LARGE, COMPRESS})
 	public byte flag;
